Creacion de las Rutas Product

+++ src/middlewares/users/deleteUser.js
@@ -4,6 +4,9 @@ const deleteUser = async (req, res, next) => {
     try{
         const id = req.body.id
         const usuario = await Usuario.findOne({where:{id}})
+        if(!usuario){
+            throw new Error(`No existe el producto con el ID: ${id}`)
+        }
         if(usuario.id > 0){
             await Usuario.destroy({where:{id:usuario.id}});
             req.body.eliminado =  usuario.id;
